requirements.txt
tensorflow
numpy
Pillow
matplotlib
gradio
tensorflow_hub
